enhance:完善后端逻辑

+++ backend/src/main/java/com/review/agent/repository/AnalysisTagRepository.java
@@ -4,5 +4,5 @@ import com.review.agent.entity.AnalysisTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AnalysisTagRepository extends JpaRepository<AnalysisTag, Long> {
-    long countByTagId(Integer tagId);
+    long countByTagId(Long tagId);
 }
